Arranging folder structure to make more sense. Currently the GM Guide is not out so the Creature/Monster/NPC folders will be hashed out more once it's released.
The three "modes" of play, Encounter/Exploration/Downtime now have a folderfor their respective log

Namespaces have been updated according to the folder structure.
