+++ pf2-core/Core/Rules/Encounters/ActionCost.cs
@@ -1,4 +1,4 @@
-namespace PF2.Core.Rules.Combat
+namespace PF2.Core.Rules.Encounters
 {
     public class ActionCost
     {
