+++ pf2-core/Core/Rules/Encounters/Combat/ArmorClass.cs
@@ -1,4 +1,4 @@
-namespace PF2.Core.Rules
+namespace PF2.Core.Rules.Encounters.Combat
 {
     public class ArmorClass
     {
