+++ pf2-core/Core/Rules/Encounters/Combat/MeleeAttack.cs
@@ -2,7 +2,7 @@ using PF2.Core.Rules.Creature;
 using PF2.Core.Utility;
 using System.Collections.Generic;
 
-namespace PF2.Core.Rules.Combat
+namespace PF2.Core.Rules.Encounters.Combat
 {
     // Core Rulebook pg12
     public class MeleeAttack : ICheck
