+++ pf2-core/Core/Rules/Equipment/Bulk.cs
@@ -1,0 +1,12 @@
+﻿namespace PF2.Core.Rules.Equipment
+{
+    internal class Bulk
+    {
+        private double BulkAmount;
+
+        public Bulk(double value = 0)
+
+        {
+        }
+    }
+}
