+++ pf2-core/Core/Rules/Equipment/IEquipment.cs
@@ -1,9 +1,9 @@
 namespace PF2.Core.Rules.Equipment
 {
-    public interface IPfEquipment
+    public interface IEquipment
     {
         int Level { get; set; }
         Price Price { get; set; }
-        int Bulk { get; set; }
+        Bulk Bulk { get; set; }
     }
 }
